+++ AngularClient/src/main/project/element-manager.component.ts
@@ -78,7 +78,6 @@ export class ElementManagerComponent implements OnInit {
 
       if (this.selection.selected.length > 0) {
 
-        this.elementDataSource.data = null;
         this.selection.selected.forEach(element => {
           this.projectService.removeElement(element);
         });
@@ -87,7 +86,7 @@ export class ElementManagerComponent implements OnInit {
           finalize(() => {
             this.elementDataSource.data = this.project.ElementSet;
           })).subscribe();
-      }
+        }
     });
   }